+++ CHANGELOG.md
@@ -8,6 +8,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Added
+- Flagged field mapping
+
+### Added
 - Publish deleted proposals
 
 ## [0.3.1] - 2023-12-04